Renamed test_impl.rs, which was misleading.
