+++ wf_account/src/lib.rs
@@ -1,3 +1,3 @@
 pub mod account;
 pub mod error;
-pub mod test_impl;
+pub mod whiteflag_account;
